Added infrastructure to disable introspection
As for Meson 0.43.0, no way to prevent a generated
GIR file to be installed in the system, even
if disable-introspection=true is set.

May this is not an issue as libgee generates
header, VAPI and GIR, all of them needed when
developing.

Just in case g-ir-compiler is not found
no typelib is generated and so introspection
is incomplete any way. g-ir-compiler is
installed with gobject-introspection package
in almost all distributions.